requirements.txt
# SwasthyaSahayak ML Dependencies
# Python 3.10+

# Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
sentence-transformers>=2.2.2

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# NLP & Language Processing
indic-nlp-library>=0.91
sacremoses>=0.0.53
sentencepiece>=0.1.99

# Vector Database & Retrieval
psycopg2-binary>=2.9.0
pgvector>=0.2.0

# API & Inference Service
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0

# ML Training & Tracking
pytorch-lightning>=2.1.0
wandb>=0.16.0
tensorboard>=2.15.0

# Evaluation & Metrics
rouge-score>=0.1.2
nltk>=3.8
bert-score>=0.3.13

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
loguru>=0.7.2
pyyaml>=6.0.1

# Code Quality
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
pytest>=7.4.3
pytest-cov>=4.1.0

# Jupyter & Visualization
jupyter>=1.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

